requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2
aiofiles==24.1.0
python-dotenv==1.0.1
python-dateutil==2.9.0
